requirements.txt
Django
django-crispy-forms
crispy-bootstrap5
scipy
